Move DBAL adapters to package
